INFUND-6785  Modified the data service to use the new data-service-commons project.
The major changes are due to removing the Application and Organisation objects from the ProcessRole (referenced in the User class) and replacing them with keys.  These objects must now be fetched on demand from their repositories in the service layer.  Similarly for the User’s Profile.  Note that the tests are broken at this time.


Former-commit-id: 272a289ab2918a56f8d62c10ab00a788b9fd81c9