IFS-231-Comp-Ques-Id-Mismatch
The code seemed to be fine. It checks if the competition exists or not
and throws 404 if it does not exist. Also correctly checks whether the
question exists or not.

However the issue was some competition data was not correctly
handled/checked resulting in 500 error for existing competitions.

This has been identified and fixed now.

Change-Id: I2d1c738682902a43dc50ea4bd906fcd000af1b95