IFS-1961 - Add invite/organisation matching except for organisations that are already attached to an application
During collaborator invite, when the user confirms their Organisation,
the Organisation should be matched (i.e. checked to see if it exists in
the database already).

If it matches, it should only be linked to an InviteOrganisation if the
Organisation is not already attached to the application (via some other
InviteOrganisation).