IFS-2564 - Map `FormInput#allowedFileTypes` using `FileTypeCategories#displayName`
Unfortunately we are currently using the `displayName` property in the
database instead of a `toString` on the FileTypeCategories enum. Some future
work could be performed to clean this up to use just the string enumeration
instead. In the same vein, we should also evaluate if we want to use a formal
many-to-many relationship between FileTypeCategories and FormInput in the
database.