IFS-2394 - Move data-service SecurityConfig out of data-service-commons
It is currently problematic to have this kind of config inside of the
commons module as it gets propagated down to the finance-data-service due
to it's `@Configuration` annotation. This means it makes incorrect assertions
about how the finance-data-service should be configured.

This highlights problems with how we've currently structured our project
files. This should be assessed further in some other tech debt ticket.